+++ apps/dashboard/next.config.ts
@@ -2,7 +2,12 @@ import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
   /* config options here */
-  transpilePackages: ['@bubbles/ui', '@bubbles/theme'],
+  transpilePackages: [
+    '@bubbles/markdown-editor',
+    '@bubbles/markdown-renderer',
+    '@bubbles/theme',
+    '@bubbles/ui',
+  ],
   reactCompiler: true,
 };