+++ apps/dashboard/vitest.config.ts
@@ -1,7 +1,7 @@
-import path from 'path'
+import path from 'path';
 
-import react from '@vitejs/plugin-react'
-import { defineConfig } from 'vitest/config'
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   plugins: [react()],
@@ -16,4 +16,4 @@ export default defineConfig({
       '@': path.resolve(__dirname, './'),
     },
   },
-})
+});